Added Health UI, Enemy Attack
Added a health slider that changes in accordance when the player takes damage.
Added scripts: HealthBehavior, HealthBarBehavior, EnemyAttackBehavior.

Health and HealthBar behavior handle taking damage and displaying it to the health bar.

EnemyAttack, when the enemy collides with the player subtracts from the players health according to the damage set.